Fundooo/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.1.0
boto3==1.10.45
botocore==1.13.45
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
CodeConvert==3.0.2
cryptography==2.8
Django==3.0.1
django-admin==2.0.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-colorful==1.3
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-excel-response2==2.0.8
django-filter==2.2.0
django-model-utils==4.0.0
django-models-ext==1.1.9
django-pandas==0.6.1
django-redis==4.11.0
django-s3direct==1.1.5
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.8
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
furl==2.1.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.4.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
numpy==1.18.1
orderedmultidict==1.0.1
pandas==1.0.0
pathtools==0.1.2
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
pyee==6.0.0
PyJWT==1.7.1
pymitter==0.2.3
python-crontab==2.4.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
s3transfer==0.2.1
screen==1.0.1
shortuuid==0.5.0
six==1.13.0
sqlparse==0.3.0
supervisor==4.1.0
TimeConvert==1.5.1
tzlocal==2.0.0
urllib3==1.25.7
vine==1.3.0
watchdog==0.10.1
xlwt==1.3.0
zipp==2.0.1
